# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdomain-prep-for-pgdp",
    "description": "Convert a folder/zip of scanned book images into a PGDP-ready submission package.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/projects/{project_id}/run-dirty": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Project Run Dirty",
        "description": "Submit a project_run_dirty job.\n\nFans out across every page in the project, running every dirty or\nnot-run stage in DAG order.  An optional ``stage_filter`` query\nparameter restricts the sweep to a single stage_id.",
        "operationId": "project_run_dirty",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "stage_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage Filter"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/projects/{project_id}/pages/{idx0}/stages/{stage_id}/run": {
      "post": {
        "tags": [
          "pages"
        ],
        "summary": "Run Page Stage",
        "description": "Run one stage on one page synchronously and return the new row.\n\nSpec: `docs/specs/pipeline-task-model.md` \u00a7\"Per-page stage runner\"\n+ \u00a7\"API surface\". Slice 4 ships the synchronous path for the simple\nimage-processing stages (grayscale/threshold/invert today; more land\nstage-by-stage). When slow stages (`ocr`, `extract_illustrations`)\nget wired, this route gains an optional `?async=true` that returns a\nJob id instead \u2014 the chip rail will poll the job's status.\n\nStatus codes:\n\n- 200: stage ran cleanly; body is the freshly-committed PageStageState.\n- 404: project not found, page not found, or cross-user access (the\n  404-not-403 pattern matches the list endpoint and avoids leaking\n  project existence).\n- 422: unknown `stage_id` (validated against PAGE_STAGE_IDS).\n- 409 Conflict: the stage's `depends_on` rows aren't all `clean`.\n  Body names the missing parents so the UI can prompt the user to\n  run them first.\n- 501 Not Implemented: the stage emits a compound output (`ocr`,\n  `extract_illustrations`, `text_review`) and the multi-artifact\n  writer hasn't shipped yet. Body has a clear \"queued for a future\n  slice\" message.\n- 500: the registered stage impl raised, OR the dual-write commit\n  failed. The page_stages row is already marked `failed` with the\n  error_message \u2014 the chip rail's next refresh will show the\n  failure inline.",
        "operationId": "run_page_stage",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "idx0",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idx0"
            }
          },
          {
            "name": "stage_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stage Id"
            }
          },
          {
            "name": "async",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Async"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Job"
                    },
                    {
                      "$ref": "#/components/schemas/PageStageState"
                    }
                  ],
                  "title": "Response Run Page Stage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Recent Jobs",
        "operationId": "list_recent_jobs",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Recent Jobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gpu/jobs": {
      "get": {
        "tags": [
          "gpu"
        ],
        "summary": "List Jobs",
        "description": "List the most recent jobs for the current user (newest first).",
        "operationId": "list_gpu_jobs",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Gpu Jobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/jobs/{job_id}": {
      "get": {
        "tags": [
          "gpu"
        ],
        "summary": "Get Job",
        "operationId": "get_gpu_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gpu"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_gpu_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Retry Job",
        "description": "Create a fresh copy of a failed/cancelled job in `queued` status.\n\nSame project_id and type \u2014 fresh id and timestamps. The original job\nstays in the database so the user can compare. The new job's payload\nstarts as a copy of the original's; if `body.payload_override` is\nprovided (P3 #16), its keys are shallow-merged over the original\npayload (override keys replace, others are preserved). Pass `None` /\nomit the body to retry verbatim.",
        "operationId": "retry_gpu_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetryJobRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "gpu"
        ],
        "summary": "Job Events",
        "description": "SSE \u2014 push status transitions until the job is terminal.\n\nSubscribes to the in-process `JobEventBroker` and forwards events as\n`text/event-stream` frames. First frame is the current snapshot from\nthe database so a late subscriber sees state immediately; subsequent\nframes come from the broker (zero-poll).",
        "operationId": "stream_gpu_job_events",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "estimated_pages": {
            "type": "integer",
            "title": "Estimated Pages",
            "default": 0
          },
          "dispatch_mode": {
            "type": "string",
            "enum": [
              "immediate",
              "scheduled"
            ],
            "title": "Dispatch Mode",
            "default": "immediate"
          },
          "next_dispatch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Dispatch At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "BatchJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "next_dispatch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Dispatch At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "gpu_backend": {
            "type": "string",
            "enum": [
              "local",
              "cpu",
              "modal",
              "shared_container"
            ],
            "title": "Gpu Backend",
            "default": "local"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "owner_id",
          "type",
          "status",
          "progress",
          "created_at",
          "started_at",
          "completed_at",
          "next_dispatch_at",
          "error_message",
          "gpu_backend",
          "payload"
        ],
        "title": "Job"
      },
...
      "JobSubmitResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobSubmitResponse",
        "description": "Minimal response for project-level job submission routes."
      },
...
      "PageStageState": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "page_id": {
            "type": "string",
            "title": "Page Id"
          },
          "stage_id": {
            "type": "string",
            "title": "Stage Id"
          },
          "status": {
            "$ref": "#/components/schemas/PageStageStatus",
            "default": "not-run"
          },
          "stage_version": {
            "type": "integer",
            "title": "Stage Version",
            "default": 1
          },
          "artifact_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Key"
          },
          "config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Hash"
          },
          "input_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Hash"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "page_id",
          "stage_id",
          "status",
          "stage_version",
          "artifact_key",
          "config_hash",
          "input_hash",
          "last_run_at",
          "duration_ms",
          "error_message",
          "job_id"
        ],
        "title": "PageStageState",
        "description": "One per-page stage row \u2014 state of stage `stage_id` on page `page_id`.\n\nSpec: `docs/specs/pipeline-task-model.md` \u00a7\"SQLite schema\" (Q1 lock).\n\n`page_id` is a string so it can encode split-child identity later (e.g.\n`0042/splits/a` for the first split-child of page 42). For root pages\ntoday, `page_id` is the zero-padded 4-digit `idx0`."
      },
...
      "PipelineState": {
        "properties": {
          "steps": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StepState"
            },
            "type": "object",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "PipelineState"
      },
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "proof_page_count": {
            "type": "integer",
            "title": "Proof Page Count"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          },
          "storage_prefix": {
            "type": "string",
            "title": "Storage Prefix"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "stage_artifacts_bytes": {
            "type": "integer",
            "title": "Stage Artifacts Bytes",
            "default": 0
          },
          "source_zip_bytes": {
            "type": "integer",
            "title": "Source Zip Bytes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "created_at",
          "updated_at",
          "status",
          "page_count",
          "proof_page_count",
          "config",
          "pipeline_state",
          "storage_prefix",
          "archived",
          "stage_artifacts_bytes",
          "source_zip_bytes"
        ],
        "title": "Project"
      },
...
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "ingesting",
          "configuring",
          "processing",
          "reviewing",
          "packaging",
          "complete"
        ],
        "title": "ProjectStatus"
      },
...
      "RetryJobRequest": {
        "properties": {
          "payload_override": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Override"
          }
        },
        "type": "object",
        "title": "RetryJobRequest",
        "description": "Optional body for `POST /api/gpu/jobs/{id}/retry`.\n\n`payload_override`, when non-null, is shallow-merged over the original\njob's payload \u2014 keys present in the override replace the corresponding\nkeys; keys not present are preserved from the original. The original\njob is never mutated, so the audit trail stays intact."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}